other process

+++ tests/test.rs
@@ -52,7 +52,10 @@ async fn page(which: Which) {
     let persistent = browser_context::persistent(&browser_type, port, which).await;
     let browser_context = browser_context::all(&browser, &persistent, which).await;
     page::all(&browser_context, port, which).await;
-    std::fs::remove_dir_all(temp_dir()).unwrap();
+    if !cfg!(windows) {
+        // error because other process uses on windows
+        std::fs::remove_dir_all(temp_dir()).unwrap();
+    }
 }
 
 async fn selectors(which: Which) {
